Refactor screenshot code into separate class.
This will allow future alternate implementations.
